robots.txt: allow more indexing
A long time ago, web crawlers were discouraged from using SSL to crawl
bugzilla, because of CPU load caused in SSL, and that it was a duplicate
of results on the non-SSL side and the bots got that wrong.

In the SSL-Everywhere world, with forced upgrades to SSL, and cheaper
CPU, this was never revisited, and means that Gentoo Bugzilla isn't on
Google. Re-enable the safer terms, and let the bots come in!

Fixes: https://bugs.gentoo.org/show_bug.cgi?id=690438
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>